skip species selection if no modules requiring species selected, updated required species error message
